Fix dependencies
The dependencies required by CImg are configured by defining macros.

Define cimg_display=0 so that the X11 library is not required.

Define cimg_use_jpeg and cimg_use_png so that the JPEG and PNG
libraries will be used, instead of shelling out to ImageMagick. This
should fix the errors mentioned in the README.